+++ src/components/App/App.tsx
@@ -1,7 +1,7 @@
 //external modules
 import React from "react";
 
-import Navigation from "../Nav";
+import Navigation from "../Navigation";
 
 const App: React.FC = () => {
   return <Navigation />;
