+++ src/components/Navigation/tests/NavigationComponent.test.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import configureMockStore from "redux-mock-store";
 import { defaultState } from "../../App/DefaultState";
 import { Provider } from "react-redux";
-import Navigation from "../";
+import Navigation from "..";
 
 const createMockStore = configureMockStore([]);
